requirements.txt
coverage==4.1
Django==1.9.7
django-nose==1.4.4
django-twitterbootstrap-form==0.4.0
django-widget-tweaks==1.4.1
factory-boy==2.7.0
fake-factory==0.5.8
ipaddress==1.0.16
nose==1.3.7
python-dateutil==2.5.3
six==1.10.0
